+++ tests/testthat.R
@@ -1,3 +1,7 @@
+# https://github.com/Rdatatable/data.table/issues/5658
+Sys.setenv("OMP_THREAD_LIMIT" = 2)
+Sys.setenv("Ncpu" = 2)
+
 library(testthat)
 library(DQAstats)
 library(utils)
@@ -5,7 +9,4 @@ library(utils)
 if (getRversion() >= "2.15.1") {
   utils::globalVariables(c("logfile_dir"))
 }
-# https://github.com/Rdatatable/data.table/issues/5658
-Sys.setenv("OMP_THREAD_LIMIT" = 2)
-Sys.setenv("Ncpu" = 2)
 test_check("DQAstats")