+++ tests/testthat/test-descriptive_results.R
@@ -43,9 +43,6 @@ test_that("correct functioning of descriptive results", {
   # set headless (without GUI, progressbars, etc.)
   rv$headless <- TRUE
 
-  # test sequentially
-  rv$parallel <- FALSE
-
   # get configs
   rv$source$settings <- DIZutils::get_config_env(
     system_name = rv$source$system_name,
@@ -190,9 +187,6 @@ test_that("correct functioning of descriptive results - single source", {
 
   # set headless (without GUI, progressbars, etc.)
   rv$headless <- TRUE
-
-  # test sequentially
-  rv$parallel <- FALSE
 
   # get configs
   rv$source$settings <- DIZutils::get_config_env(